move it back! update copy!
